requirements.txt
fastapi
uvicorn
supabase
anthropic
python-dotenv
httpx
apscheduler
apify-client
streamlit
